requirements.txt
streamlit==1.15.0
torch==1.12.1
Pillow==9.0.1
gdown==4.6.0
